requirements.txt
web3>=6.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
# onchainos CLI: curl -sSL https://raw.githubusercontent.com/okx/onchainos-skills/main/install.sh | sh
